refactor(websocket): extract CommandRouter from the runner
T6 of docs/websocket-runner-refactor-plan.md. The 540-line handleCommand
switch was the last large region left on the runner, and it reached across
every other one: job control, chat turns, direct inference, tRPC, the socket
itself. Nothing documented which of the runner's fields a command was allowed
to touch, because all of them were in scope.

The switch is now a dispatch table on CommandRouter, which reaches the other
regions through their own methods rather than their state. Two commands that
read job state directly did so through the job map: `stream_input` logged its
keys and `stop` cancelled an entry in it, so JobExecutionManager grew
`activeJobIds()` and `stopJob()` and the map stays private to it.

What a command does to the connection itself — the wire mode, the RPC abort
set, the tool bridges, the chat-turn session bookkeeping the frame router
reads — stays host state, reached through an adapter the runner passes in.
The runner keeps `handleCommand` and `runRpc` as delegators onto the router:
the wire entry point and the suites that drive the RPC frame are unchanged.

`currentTask` goes with the switch. Its only read was the `void
this.currentTask.catch(...)` chain in the `inference` case, which is a local
now; nothing else ever read the field.

The SDK v1 execution contract test greps live source for one dispatch site per
declared command. Its pointer moves to the file the dispatch moved to, and the
assertion keeps its strength — verified by renaming a key and watching it fail.
